[ADDED] pack label
Former-commit-id: e257e2e0494df46b253d4099217da50ecf4e8cab

+++ easysprpacker/src/libsprpacker/PackLabel.cpp
@@ -79,7 +79,7 @@ int PackLabel::SizeOfPackToBin() const
 {
 	int sz = 0;
 
-	sz += sizeof(uint16_t);				// id
+	sz += sizeof(uint32_t);				// id
 	sz += sizeof(uint8_t);				// type
 
 	sz += sizeof(uint16_t) * 2;			// width, height
